test: add test page component
